+++ Color/src/Graphics/Color/Space/CIE1976/LAB.hs
@@ -24,6 +24,13 @@ module Graphics.Color.Space.CIE1976.LAB
   , pattern ColorLAB
   , pattern ColorLABA
   , LAB
+    -- * Helpers
+    -- ** XYZ to L*a*b*
+  , xyz2lab
+  , ft
+    -- ** L*a*b* to XYZ
+  , lab2xyz
+  , ift
   ) where
 
 import Data.List.NonEmpty
@@ -36,6 +43,52 @@ import Graphics.Color.Space.Internal
 --------------
 
 -- | [CIE L*a*b*](https://en.wikipedia.org/wiki/CIELAB_color_space) color space
+--
+-- It is customary to have CIELAB color channels to be in range of [0, 100], however in
+-- this library all values for consistency are kept in a [0, 1] range for floating point
+-- precision.
+--
+-- Conversion from `XYZ` (`xyz2lab`):
+--
+-- \[
+-- \begin{align}
+--   L^\star &= 1.16 \  f\!\left(\frac{Y}{Y_{\mathrm{n}}}\right) - 0.16\\
+--   a^\star &= 5.0 \left(f\!\left(\frac{X}{X_{\mathrm{n}}}\right) - f\!\left(\frac{Y}{Y_{\mathrm{n}}}\right)\right)\\
+--   b^\star &= 2.0 \left(f\!\left(\frac{Y}{Y_{\mathrm{n}}}\right) - f\!\left(\frac{Z}{Z_{\mathrm{n}}}\right)\right)
+-- \end{align}
+-- \]
+--
+-- Where `ft` is defined as:
+--
+-- \[
+-- \begin{align}
+--   f(t) &= \begin{cases}
+--     \sqrt[3]{t} & \text{if } t > \delta^3 \\
+--     \dfrac{t}{3 \delta^2} + \frac{4}{29} & \text{otherwise}
+--   \end{cases} \\
+--   \delta &= \tfrac{6}{29}
+-- \end{align}
+-- \]
+--
+-- Conversion to `XYZ` (`lab2xyz`):
+--
+-- \[
+-- \begin{align}
+--   X &= X_{\mathrm{n}} f^{-1}\left(\frac{L^\star+0.16}{1.16} + \frac{a^\star}{5.0}\right)\\
+--   Y &= Y_{\mathrm{n}} f^{-1}\left(\frac{L^\star+0.16}{1.16}\right)\\
+--   Z &= Z_{\mathrm{n}} f^{-1}\left(\frac{L^\star+0.16}{1.16} - \frac{b^\star}{2.0}\right)\\
+-- \end{align}
+-- \]
+--
+-- Where `ift` is defined as:
+--
+-- \[
+-- f^{-1}(t) = \begin{cases}
+--   t^3 & \text{if } t > \delta \\
+--   3\delta^2\left(t - \tfrac{4}{29}\right) & \text{otherwise}
+-- \end{cases}
+-- \]
+--
 data LAB (i :: k)
 
 -- | Color in CIE L*a*b* color space
@@ -116,22 +169,22 @@ lab2xyz ::
   -> Color (XYZ i) a
 lab2xyz (ColorLAB l' a' b') = ColorXYZ x y z
   where
-    !(ColorXYZ wx _ wz) = whitePointTristimulus :: Color (XYZ i) a
+    ColorXYZ wx _ wz = whitePointTristimulus :: Color (XYZ i) a
     !l = scaleLightness l'
-    !x = wx * ift (l + toRealFloat a' / 500)
+    !x = wx * ift (l + toRealFloat a' / 5)
     !y = ift l
-    !z = wz * ift (l - toRealFloat b' / 200)
+    !z = wz * ift (l - toRealFloat b' / 2)
 {-# INLINE lab2xyz #-}
 
 scaleLightness :: (Elevator e, Elevator a, RealFloat a) => e -> a
-scaleLightness l' = (toRealFloat l' + 16) / 116
+scaleLightness l' = (toRealFloat l' + 0.16) / 1.16
 {-# INLINE scaleLightness #-}
 
 ift :: (Fractional a, Ord a) => a -> a
 ift t
   | t > 6 / 29 = t ^ (3 :: Int)
   | otherwise = (108 / 841) * (t - 4 / 29)
-
+{-# INLINE ift #-}
 
 
 xyz2lab ::
@@ -140,13 +193,13 @@ xyz2lab ::
   -> Color (LAB i) e
 xyz2lab (ColorXYZ x y z) = ColorLAB l' a' b'
   where
-    !(ColorXYZ wx _ wz) = whitePointTristimulus :: Color (XYZ i) e
+    ColorXYZ wx _ wz = whitePointTristimulus :: Color (XYZ i) e
     !fx = ft (toRealFloat x / wx)
     !fy = ft (toRealFloat y)
     !fz = ft (toRealFloat z / wz)
-    !l' = 116 * fy - 16
-    !a' = 500 * (fx - fy)
-    !b' = 200 * (fy - fz)
+    !l' = 1.16 * fy - 0.16
+    !a' = 5 * (fx - fy)
+    !b' = 2 * (fy - fz)
 {-# INLINE xyz2lab #-}
 
 ft :: RealFloat a => a -> a